New underline view in progress. Cosmetics on stage bar.

+++ MobiUs/FSTDottedDividerView.m
@@ -17,7 +17,7 @@
     
     //dotted line down the center
     dottedPath.lineWidth = 0.8;
-    float dashPattern[] = {2, 4};
+    CGFloat dashPattern[] = {2, 4};
     [dottedPath setLineDash:dashPattern count:2 phase:0];
     [dottedPath moveToPoint:CGPointMake(rect.size.width/2, rect.origin.y)];
     [dottedPath addLineToPoint:CGPointMake(rect.size.width/2, rect.origin.y = rect.size.height)];
